backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3
google-generativeai==0.3.2
aiofiles==23.2.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.9.1
email-validator==2.1.0.post1
